+++ composite.go
@@ -1,17 +1,17 @@
 package emperror
 
-// CompositeHandler allows an error to be processed by multiple handlers.
-type CompositeHandler struct {
+// compositeHandler allows an error to be processed by multiple handlers.
+type compositeHandler struct {
 	handlers []Handler
 }
 
-// NewCompositeHandler returns a new CompositeHandler.
+// NewCompositeHandler returns a new compositeHandler.
 func NewCompositeHandler(handlers ...Handler) Handler {
-	return &CompositeHandler{handlers}
+	return &compositeHandler{handlers}
 }
 
 // Handle goes through the handlers and call each of them for the error.
-func (h *CompositeHandler) Handle(err error) {
+func (h *compositeHandler) Handle(err error) {
 	for _, handler := range h.handlers {
 		handler.Handle(err)
 	}