+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Changed
+
+- `compositeHandler` not exported anymore
+
+
 ## 0.3.0 - 2017-07-11
 
 ### Added